+++ src/converters/piral-ng/src/common/piral-ng-common.js
@@ -1,0 +1,3 @@
+exports.NgExtension = class {};
+exports.SharedModule = class {};
+exports.ResourceUrlPipe = class {};