+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Piral Changelog
 
+## 0.15.2 (tbd)
+
+- Fixed import of common module in CommonJS build of `piral-ng`
+- Fixed issue of Angular components not being disposed (#561)
+- Improved loading of satellite dependencies in `piral-blazor`
+- Added `piral run-emulator` command to `piral-cli`
+- Added support for capabilities to `piral-blazor`
+
 ## 0.15.1 (November 25, 2022)
 
 - Fixed update of `piral-extension` web component inside foreign components